修正行走时间
