+++ assets/script/utils/get-random-int.ts
@@ -1,8 +1,8 @@
 /**
- * Get a random integer between `min` and `max`.
+ * 获取在 `min` 和 `max`之间的随机整数
  *
- * @param {number} min - min number
- * @param {number} max - max number
+ * @param {number} min - 最小数
+ * @param {number} max - 最大数
  * @return {number} a random integer
  */
 export function getRandomInt(min: number, max: number): number {
